Add support for arrays and tuples in smock
