+++ src/smockit/index.ts
@@ -1,0 +1,2 @@
+export * from './smockit'
+export * from './types'
