+++ src/smoddit/index.ts
@@ -1,0 +1,2 @@
+export * from './smoddit'
+export * from './types'
