+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './hex-utils'
+export * from './address-utils'
